Chatbot/first-chatbot-backend/requirements.txt
anyio==4.2.0
asgiref==3.7.2
bardapi==0.1.39
certifi==2024.2.2
chardet==3.0.4
colorama==0.4.6
Cython==0.29.14
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==2.8
interruptingcow==0.8
nltk==3.4.5
psutil==5.6.7
pygtrie==2.3.2
pyjnius==1.2.0
pyparsing==2.4.5
pytz==2024.1
pywin32==227
requests==2.22.0
six==1.13.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.4
Unidecode==1.1.1
urllib3==1.25.11
vlibras-translate==1.1.1
